+++ crates/swc_ecma_minifier/src/compress/optimize/util.rs
@@ -1,4 +1,7 @@
-use std::ops::{Deref, DerefMut};
+use std::{
+    mem::take,
+    ops::{Deref, DerefMut},
+};
 
 use rustc_hash::{FxHashMap, FxHashSet};
 use swc_atoms::js_word;
@@ -528,6 +531,45 @@ impl VisitMut for ExprReplacer {
                     key: PropName::Ident(i.clone()),
                     value,
                 });
+            }
+        }
+    }
+}
+
+#[derive(Debug, Default, PartialEq, Eq)]
+pub(super) struct SynthesizedStmts(Vec<Stmt>);
+
+impl SynthesizedStmts {
+    pub fn take_stmts(&mut self) -> Vec<Stmt> {
+        take(&mut self.0)
+    }
+}
+
+impl std::ops::Deref for SynthesizedStmts {
+    type Target = Vec<Stmt>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl std::ops::DerefMut for SynthesizedStmts {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.0
+    }
+}
+
+impl Take for SynthesizedStmts {
+    fn dummy() -> Self {
+        Self(Take::dummy())
+    }
+}
+
+impl Drop for SynthesizedStmts {
+    fn drop(&mut self) {
+        if !self.0.is_empty() {
+            if !std::thread::panicking() {
+                panic!("We should not drop synthesized stmts");
             }
         }
     }
